+++ packages/admin-ui/src/Checkbox/index.ts
@@ -1,4 +1,6 @@
 export * from "./Checkbox";
-export * from "./CheckboxGroup";
-export * from "./CheckboxGroupPresenter";
-export * from "./CheckboxPresenter";
+export * from "./CheckboxItem";
+export * from "./CheckboxItemDto";
+export * from "./CheckboxItemFormatted";
+export * from "./CheckboxItemMapper";
+export * from "./CheckboxPrimitive";
