+++ packages/admin-ui/src/CheckboxGroup/index.ts
@@ -1,0 +1,2 @@
+export * from "./CheckboxGroup";
+export * from "./CheckboxGroupPrimitive";
