--- packages/admin-ui/src/Form/Checkbox/index.ts
@@ -1,2 +1,0 @@
-export * from "./Checkbox";
-export * from "./CheckboxGroup";
