--- packages/admin-ui/src/Form/Input/index.ts
@@ -1,1 +1,0 @@
-export * from "./Input";
