--- packages/admin-ui/src/Form/Select/index.ts
@@ -1,1 +1,0 @@
-export * from "./Select";
