--- packages/admin-ui/src/Form/Switch/index.ts
@@ -1,1 +1,0 @@
-export * from "./Switch";
