+++ packages/admin-ui/src/Form/index.ts
@@ -1,7 +1,2 @@
-export * from "./Checkbox";
-export * from "./FormComponent";
-export * from "./Input";
 export * from "./RangeSlider";
-export * from "./Select";
 export * from "./Slider";
-export * from "./Switch";
