+++ packages/admin-ui/src/FormComponent/Label.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { makeDecoratable } from "~/utils";
+import { Label, LabelProps } from "~/Label";
+
+interface FormComponentLabelProps {
+    text?: React.ReactElement<typeof Label> | React.ReactNode;
+    required?: boolean;
+}
+
+const DecoratableFormComponentLabel = (props: FormComponentLabelProps) => {
+    if (!props.text) {
+        return null;
+    }
+
+    if (React.isValidElement(props.text) && props.text.type === Label) {
+        return React.cloneElement(props.text as React.ReactElement<LabelProps>, {
+            required: props.required
+        });
+    }
+
+    return <Label text={props.text} required={props.required} />;
+};
+
+const FormComponentLabel = makeDecoratable("FormComponentLabel", DecoratableFormComponentLabel);
+
+export { FormComponentLabel, type FormComponentLabelProps };
