+++ packages/admin-ui/src/FormComponent/index.ts
@@ -1,0 +1,1 @@
+export * from "./FormComponent";
