+++ packages/admin-ui/src/Select/Select.stories.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { ReactComponent as SearchIcon } from "@material-design-icons/svg/outlined/search.svg";
-import { ReactComponent as ChevronRight } from "@material-design-icons/svg/outlined/chevron_right.svg";
-import { Select } from "./Select";
-import { Button } from "~/Button";
+import { ReactComponent as NotificationsIcon } from "@material-design-icons/svg/outlined/notifications.svg";
+import { ReactComponent as CalendarIcon } from "@material-design-icons/svg/outlined/calendar_month.svg";
+import { Select } from "~/Select";
+import { Icon } from "~/Icon";
 
 const meta: Meta<typeof Select> = {
     title: "Components/Select",
@@ -11,11 +11,7 @@ const meta: Meta<typeof Select> = {
     tags: ["autodocs"],
     argTypes: {
         onValueChange: { action: "onValueChange" },
-        onOpenChange: { action: "onOpenChange" },
-        variant: { control: "select", options: ["primary", "secondary", "ghost"] },
-        size: { control: "select", options: ["md", "lg", "xl"] },
-        disabled: { control: "boolean" },
-        invalid: { control: "boolean" }
+        onOpenChange: { action: "onOpenChange" }
     },
     parameters: {
         layout: "fullscreen"
@@ -23,7 +19,9 @@ const meta: Meta<typeof Select> = {
     decorators: [
         Story => (
             <div className="w-1/3 h-64 mx-auto flex justify-center items-center">
-                <Story />
+                <div className="w-full">
+                    <Story />
+                </div>
             </div>
         )
     ],
@@ -34,7 +32,6 @@ const meta: Meta<typeof Select> = {
 };
 
 export default meta;
-
 type Story = StoryObj<typeof Select>;
 
 export const Default: Story = {
@@ -59,270 +56,67 @@ export const Default: Story = {
     }
 };
 
-export const MediumSize: Story = {
+export const WithLabel: Story = {
     args: {
         ...Default.args,
-        size: "md"
-    }
-};
-
-export const LargeSize: Story = {
-    args: {
-        ...Default.args,
-        size: "lg"
-    }
-};
-
-export const ExtraLargeSize: Story = {
-    args: {
-        ...Default.args,
-        size: "xl"
-    }
-};
-
-export const WithCustomPlaceholder: Story = {
-    args: {
-        ...Default.args,
-        placeholder: "Custom placeholder"
+        label: "Timezone"
     }
 };
 
 export const WithStartIcon: Story = {
     args: {
         ...Default.args,
-        startIcon: <SearchIcon />
+        label: "Timezone",
+        placeholder: "Pick a timezone",
+        startIcon: <Icon label={"Bell"} icon={<NotificationsIcon />} />
     }
 };
 
-export const WithEndIconIcon: Story = {
+export const WithEndIcon: Story = {
     args: {
         ...Default.args,
-        endIcon: <SearchIcon />
+        label: "Timezone",
+        placeholder: "Pick a timezone",
+        endIcon: <Icon label={"Calendar"} icon={<CalendarIcon />} />
     }
 };
 
-export const WithStartAndEndIcons: Story = {
+export const WithDescription: Story = {
     args: {
         ...Default.args,
-        startIcon: <SearchIcon />,
-        endIcon: <ChevronRight />
+        description:
+            "We need your timezone to make the press cards ready for your pickup at the conference."
     }
 };
 
-export const PrimaryVariant: Story = {
+export const WithNotes: Story = {
     args: {
         ...Default.args,
-        variant: "primary"
+        note: "Note: Please select only a valid timezone."
     }
 };
 
-export const PrimaryVariantDisabled: Story = {
-    args: {
-        ...PrimaryVariant.args,
-        disabled: true
-    }
-};
-
-export const PrimaryVariantInvalid: Story = {
-    args: {
-        ...PrimaryVariant.args,
-        invalid: true
-    }
-};
-
-export const SecondaryVariant: Story = {
-    args: {
-        variant: "secondary",
-        placeholder: "Custom placeholder"
-    }
-};
-
-export const SecondaryVariantDisabled: Story = {
-    args: {
-        ...SecondaryVariant.args,
-        disabled: true
-    }
-};
-
-export const SecondaryVariantInvalid: Story = {
-    args: {
-        ...SecondaryVariant.args,
-        invalid: true
-    }
-};
-
-export const GhostVariant: Story = {
-    args: {
-        variant: "ghost",
-        placeholder: "Custom placeholder"
-    }
-};
-
-export const GhostVariantDisabled: Story = {
-    args: {
-        ...GhostVariant.args,
-        disabled: true
-    }
-};
-
-export const GhostVariantInvalid: Story = {
-    args: {
-        ...GhostVariant.args,
-        invalid: true
-    }
-};
-
-export const WithFormattedOptions: Story = {
+export const WithErrors: Story = {
     args: {
         ...Default.args,
-        options: [
-            { label: "Eastern Standard Time (EST)", value: "est" },
-            { label: "Central Standard Time (CST)", value: "cst" },
-            { label: "Pacific Standard Time (PST)", value: "pst" },
-            { label: "Greenwich Mean Time (GMT)", value: "gmt" },
-            { label: "Central European Time (CET)", value: "cet" },
-            { label: "Central Africa Time (CAT)", value: "cat" },
-            { label: "India Standard Time (IST)", value: "ist" },
-            { label: "China Standard Time (CST)", value: "cst_china" },
-            { label: "Japan Standard Time (JST)", value: "jst" },
-            { label: "Australian Western Standard Time (AWST)", value: "awst" },
-            { label: "New Zealand Standard Time (NZST)", value: "nzst" },
-            { label: "Fiji Time (FJT)", value: "fjt" },
-            { label: "Argentina Time (ART)", value: "art" },
-            { label: "Bolivia Time (BOT)", value: "bot" },
-            { label: "Brasilia Time (BRT)", value: "brt" }
-        ]
+        validation: {
+            isValid: false,
+            message: "Numerical values are not allowed."
+        }
     }
 };
 
-export const WithOptionGroups: Story = {
+export const FullExample: Story = {
     args: {
         ...Default.args,
-        options: [
-            {
-                label: "North America",
-                options: [
-                    { label: "Eastern Standard Time (EST)", value: "est" },
-                    { label: "Central Standard Time (CST)", value: "cst" },
-                    { label: "Pacific Standard Time (PST)", value: "pst" }
-                ]
-            },
-            {
-                label: "Europe & Africa",
-                options: [
-                    { label: "Greenwich Mean Time (GMT)", value: "gmt" },
-                    { label: "Central European Time (CET)", value: "cet" },
-                    { label: "Central Africa Time (CAT)", value: "cat" }
-                ]
-            },
-            {
-                label: "Asia",
-                options: [
-                    { label: "India Standard Time (IST)", value: "ist" },
-                    { label: "China Standard Time (CST)", value: "cst_china" },
-                    { label: "Japan Standard Time (JST)", value: "jst" }
-                ]
-            },
-            {
-                label: "Australia & Pacific",
-                options: [
-                    { label: "Australian Western Standard Time (AWST)", value: "awst" },
-                    { label: "New Zealand Standard Time (NZST)", value: "nzst" },
-                    { label: "Fiji Time (FJT)", value: "fjt" }
-                ]
-            },
-            {
-                label: "South America",
-                options: [
-                    { label: "Argentina Time (ART)", value: "art" },
-                    { label: "Bolivia Time (BOT)", value: "bot" },
-                    { label: "Brasilia Time (BRT)", value: "brt" }
-                ]
-            }
-        ]
-    }
-};
-
-export const WithSeparators: Story = {
-    args: {
-        ...Default.args,
-        options: [
-            { label: "Eastern Standard Time (EST)", value: "est" },
-            { label: "Central Standard Time (CST)", value: "cst" },
-            { label: "Pacific Standard Time (PST)", value: "pst", separator: true },
-            { label: "Greenwich Mean Time (GMT)", value: "gmt" },
-            { label: "Central European Time (CET)", value: "cet" },
-            { label: "Central Africa Time (CAT)", value: "cat", separator: true },
-            { label: "India Standard Time (IST)", value: "ist" },
-            { label: "China Standard Time (CST)", value: "cst_china" },
-            { label: "Japan Standard Time (JST)", value: "jst", separator: true },
-            { label: "Australian Western Standard Time (AWST)", value: "awst" },
-            { label: "New Zealand Standard Time (NZST)", value: "nzst" },
-            { label: "Fiji Time (FJT)", value: "fjt", separator: true },
-            { label: "Argentina Time (ART)", value: "art" },
-            { label: "Bolivia Time (BOT)", value: "bot" },
-            { label: "Brasilia Time (BRT)", value: "brt" }
-        ]
-    }
-};
-
-export const WithDisabledOptions: Story = {
-    args: {
-        options: [
-            { label: "Eastern Standard Time (EST)", value: "est", disabled: true },
-            { label: "Central Standard Time (CST)", value: "cst", disabled: true },
-            { label: "Pacific Standard Time (PST)", value: "pst", disabled: true },
-            { label: "Greenwich Mean Time (GMT)", value: "gmt" },
-            { label: "Central European Time (CET)", value: "cet" },
-            { label: "Central Africa Time (CAT)", value: "cat" },
-            { label: "India Standard Time (IST)", value: "ist" },
-            { label: "China Standard Time (CST)", value: "cst_china" },
-            { label: "Japan Standard Time (JST)", value: "jst" },
-            { label: "Australian Western Standard Time (AWST)", value: "awst" },
-            { label: "New Zealand Standard Time (NZST)", value: "nzst" },
-            { label: "Fiji Time (FJT)", value: "fjt" },
-            { label: "Argentina Time (ART)", value: "art" },
-            { label: "Bolivia Time (BOT)", value: "bot" },
-            { label: "Brasilia Time (BRT)", value: "brt" }
-        ]
-    }
-};
-
-export const WithExternalValueControl: Story = {
-    args: {
-        ...Default.args,
-        options: [
-            "Eastern Standard Time (EST)",
-            "Central Standard Time (CST)",
-            "Pacific Standard Time (PST)",
-            "Greenwich Mean Time (GMT)",
-            "Central European Time (CET)",
-            "Central Africa Time (CAT)",
-            "India Standard Time (IST)",
-            "China Standard Time (CST)",
-            "Japan Standard Time (JST)",
-            "Australian Western Standard Time (AWST)",
-            "New Zealand Standard Time (NZST)",
-            "Fiji Time (FJT)",
-            "Argentina Time (ART)",
-            "Bolivia Time (BOT)",
-            "Brasilia Time (BRT)"
-        ]
-    },
-    render: args => {
-        const [value, setValue] = useState(args.value);
-        return (
-            <div className={"w-full"}>
-                <div>
-                    <Select {...args} value={value} onValueChange={value => setValue(value)} />
-                </div>
-                <div className={"mt-4 text-center"}>
-                    <Button text={"Reset"} onClick={() => setValue("")} />
-                </div>
-                <div className={"mt-4 text-center"}>
-                    Current selected value: <pre>{value}</pre>
-                </div>
-            </div>
-        );
+        label: "Timezone",
+        placeholder: "Pick a timezone",
+        description:
+            "We need your timezone to make the press cards ready for your pickup at the conference.",
+        note: "Note: Please select only a valid timezone.",
+        validation: {
+            isValid: false,
+            message: "Numerical values are not allowed."
+        }
     }
 };
