+++ packages/admin-ui/src/Switch/index.ts
@@ -1,2 +1,2 @@
 export * from "./Switch";
-export * from "./SwitchPresenter";
+export * from "./SwitchPrimitive";
