+++ packages/admin-ui/src/index.ts
@@ -3,6 +3,7 @@ export * from "./Avatar";
 export * from "./Button";
 export * from "./Card";
 export * from "./Checkbox";
+export * from "./FormComponent";
 export * from "./Grid";
 export * from "./Heading";
 export * from "./Input";
